db/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymysql==1.0.2
requests==2.26.0
passlib>=1.7.4
jose==1.0.0
fastapi-jwt-auth==0.5.0
pyjwt
pydantic
passlib[bcrypt]
